+++ src/parser/mrt/mrt_elem.rs
@@ -109,7 +109,8 @@ fn get_relevant_attributes(
 
             AttributeValue::OriginatorId(_)
             | AttributeValue::Clusters(_)
-            | AttributeValue::Development(_) => {}
+            | AttributeValue::Development(_)
+            | AttributeValue::LinkState(_) => {}
         };
     }